+++ CHANGES.rst
@@ -7,7 +7,8 @@ Changelog
 
 - Remove resolveuid serializer/deserializer for blocks: now they are integrated in plone.restapi.
   [cekk]
-
+- Add custom transformers for blocks to handle link in generic blocks.
+  [cekk]
 
 0.1.4 (2020-07-16)
 ------------------